requirements_agent.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
cython==3.0.12
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.11.0
faiss-gpu==1.7.2
fastjsonschema==2.21.1
filelock==3.18.0
flask==3.1.0
flatbuffers==25.2.10
fonttools==4.58.0
fqdn==1.5.1
fsspec==2025.5.1
h11==0.14.0
hdbscan==0.8.40
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
ijson==3.4.0
ipykernel==6.29.5
ipython==8.36.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.2
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.15
kiwisolver==1.4.8
langcodes==3.5.0
language-data==1.3.0
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
murmurhash==1.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.4.2
notebook-shim==0.2.4
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.21.1
openai==1.76.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
preshed==3.0.9
prometheus-client==0.22.0
prompt-toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pygments==2.19.1
pyjnius==1.6.1
pynndescent==0.5.13
pyparsing==3.2.3
pyserini==0.44.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytrec-eval==0.5
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
send2trash==1.8.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
thinc==8.3.6
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.0
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.3.0
typer==0.15.3
types-python-dateutil==2.9.0.20250516
typing-extensions==4.14.0rc1
typing-inspection==0.4.0
tzdata==2025.2
umap-learn==0.5.7
uri-template==1.3.0
urllib3==2.4.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.1.3
widgetsnbextension==4.0.14
wrapt==1.17.2
xformers==0.0.31.dev1033
